moved files to new folders and renamed canvas files for clarity to planets and stars
