+++ client/src/components/Canvas/Stars.tsx
