+++ client/src/images/Jupiter_rotate.png
