+++ client/src/images/Jupiter_rotate1.png
