+++ client/src/images/asteroid_belt.png
