+++ client/src/images/asteroid_belt_grey.png
